Authenticator.Vault

Description

Types for storing, serializing, accessing OTP keys. Gratuitous type-level programming here for no reason because I have issues.

Based off of https://github.com/google/google-authenticator.

Documentation

data Mode

OTP generation mode

Constructors

HOTP
Counter-based

TOTP
Time-based

data SMode :: Mode -> Type where

Singleton for Mode

Constructors

SHOTP :: SMode HOTP

STOTP :: SMode TOTP

withSMode :: Mode -> (forall m. SMode m -> r) -> r

Reify a Mode to its singleton

fromSMode :: SMode m -> Mode

Reflect a SMode to its value.

data HashAlgo

Which OTP-approved hash algorithm to use?

Constructors

HASHA1

HASHA256

HASHA512

parseAlgo :: String -> Maybe HashAlgo

Parse a hash algorithm string into the appropriate HashAlgo.

data Secret :: Mode -> Type where

A standards-compliant secret key type. Well, almost. It doesn't include configuration for the time period if it's time-based.

Constructors

Sec

Fields

:: { secAccount :: Text

, secIssuer :: Maybe Text

, secAlgo :: HashAlgo

, secDigits :: OTPDigits

, secKey :: ByteString

} -> Secret m

newtype OTPDigits

Newtype wrapper to provide Eq, Ord, Binary, and ToJSON instances. You can convert to and from this and the Int representation using OTPDigitsInt

Constructors

OTPDigits

Fields

otpDigits :: OTPDigits

pattern OTPDigitsInt :: OTPDigits -> Int

data family ModeState :: Mode -> Type

A data family consisting of the state required by each mode.

type SomeSecretState = DSum SMode (Secret :*: ModeState)

A Secret coupled with its ModeState, existentially quantified over its Mode.

newtype Vault

A list of secrets and their states, of various modes.

Constructors

Vault

Fields

vaultList :: [SomeSecretState]

_Vault :: Functor f => ([SomeSecretState] -> f [SomeSecretState]) -> Vault -> f Vault

A lens into the list of SomeSecretStates in a Vault. Should be an Iso but we don't want a lens dependency now, do we.

hotp :: Secret HOTP -> ModeState HOTP -> (Text, ModeState HOTP)

Generate an HTOP (counter-based) code, returning a modified state.

totp :: Secret TOTP -> IO Text

Generate a TOTP (time-based) code in IO for the current time.

totp_ :: Secret TOTP -> POSIXTime -> Text

(Purely) generate a TOTP (time-based) code, for a given time.

otp :: SMode m -> Secret m -> ModeState m -> IO (Text, ModeState m)

Abstract over both hotp and totp.

someSecret :: Functor f => (forall m. SMode m -> Secret m -> ModeState m -> f (ModeState m)) -> SomeSecretState -> f SomeSecretState

Some sort of RankN lens and traversal over a SomeSecret. Allows you to traverse (effectfully map) over the ModeState in a SomeSecretState, with access to the Secret as well.

With this you can implement getters and setters. It's also used by the library to update the ModeState in IO.

vaultSecrets :: Applicative f => (forall m. SMode m -> Secret m -> ModeState m -> f (ModeState m)) -> Vault -> f Vault

A RankN traversal over all of the Secrets and ModeStates in a Vault.

describeSecret :: Secret m -> Text

Print out the metadata (account name and issuer) of a Secret.

secretURI :: Parser SomeSecretState

A parser for a otpauth URI.

parseSecretURI :: String -> Either String SomeSecretState

Parse a valid otpauth URI and initialize its state.

See https://github.com/google/google-authenticator/wiki/Key-Uri-Format