Encrypted

Description

Basically provides Enc a, which abstracts over an encrypted a. Can only be read by invoking GnuPG in IO, where the user needs to provide their key to decrypt.

One main advantage is that an Enc a can be seriealized and deserialized using its Binary instance, providing type-safe deserialization into encrypted values.

Might be pulled out to an external package some day.

Documentation

data Enc a

An Enc a abstracts over a encrypted a.

Has a useful Binary instance, which allows type-safe deserialization into encrypted values.

Constructors

Enc

Fields

encBytes :: Encrypted

mkEnc :: Binary a => Ctx -> Key -> a -> IO (Enc a)

Wrap a value into an Enc, using a given GnuPG key.

overEnc :: Binary a => Ctx -> Key -> Enc a -> (a -> IO a) -> IO (Enc a)

Modify an encrypted value with a given a -> IO a function, re-encrypting it with the given GnuPG key. The decrypted value never leaves the closure.

getEnc :: Binary a => Ctx -> Enc a -> IO a

Extract a value from an Enc, using a compatible key in the GnuPG environment.

withEnc :: Binary a => Ctx -> Key -> Enc a -> (a -> IO (b, a)) -> IO (b, Enc a)

A variation of overEnc that allows the user to also return a value produced from the decrypted value. Re-encrypts the changed value using the given GnuPG key.